benchmark/template_functions/image-processing/requirements.txt
boto3
pillow
oss2